частично изменена архитектура проекта, добавлен новый фильтр
