+++ filters/ChatTypeFilter.py
@@ -4,10 +4,7 @@ from aiogram.dispatcher.filters import BoundFilter
 
 class IsGroup(BoundFilter):
     async def check(self, message: types.Message) -> bool:
-        return message.chat.type in (
-            types.ChatType.GROUP,
-            types.ChatType.SUPERGROUP
-        )
+        return message.chat.type in (types.ChatType.GROUP, types.ChatType.SUPERGROUP)
 
 
 class IsChat(BoundFilter):
