+++ handlers/groups/main_menu_group.py
@@ -5,9 +5,9 @@ from aiogram.utils.exceptions import BotBlocked
 from django.db.models import QuerySet
 
 from django_admin.bot.models import Help_content
-from django_admin.service.models import Users, ChatIDAdmission
+from django_admin.service.models import Users, ChatIDAdmission, ChatIDDirections
 from filters import IsGroup
-from loader import dp, bot, debug
+from loader import dp, bot, debugger
 from aiogram.types import Message, CallbackQuery
 
 from keyboards.inline import buttons as btn
@@ -34,8 +34,11 @@ group_menu: list = [
 # * Кнопка создания объявлений пользователям
 @dp.message_handler(IsGroup(), AdminFilter(), text="Сделать объявление абитуриентам")
 async def attention_message(message: Message):
-    await message.answer("Напишите сообщение")
-    await GroupState.attention_message.set()
+    if await db_commands.isChatExist(chat_id=message.chat.id):
+        await message.answer("Напишите сообщение")
+        await GroupState.attention_message.set()
+    else:
+        await message.answer("Группа не числится в базе данных")
 
 
 @dp.message_handler(state=GroupState.attention_message)
@@ -69,11 +72,13 @@ async def del_chat(message: Message):
         await message.answer("Группа удалена из базы данных")
 
     except ObjectDoesNotExist as error:
-        await message.answer(f"Группа уже удалена из базы данных или не состоит в ней вовсе.\n{debug(str(error))}")
+        await message.answer(
+            f"Группа уже удалена из базы данных или не состоит в ней вовсе.\n{await debugger(str(error))}"
+        )
         logging.error(error)
 
 
-# * Обработчик кнопки Занести группу в базу данных
+# * Обработчик кнопки "Занести группу в базу данных"
 @dp.message_handler(IsGroup(), AdminFilter(), text="Занести группу в базу данных")
 async def add_group(message: Message):
     chat_exists: bool = await db_commands.isChatExist(message.chat.id)
@@ -92,9 +97,9 @@ async def type_of_group(call: CallbackQuery, state: FSMContext):
             await call.message.edit_text("Группа успешно занесена в базу данных")
 
         except Exception as error:
-            await call.message.edit_text(f"Группа уже занесена в базу данных", parse_mode='')
-            logging.error(error)
+            await call.message.edit_text("Группа уже занесена в базу данных", parse_mode='')
 
+            logging.error(error)
     elif call.data == "dp_type":
         await call.message.edit_text("Выберите уровень подготовки", reply_markup=btn.choose_level)
 
@@ -113,14 +118,14 @@ async def ask_to_question_handler(call: CallbackQuery, state: FSMContext):
     """
     await state.set_state(Questions.answer)
 
-    callback_data = group_btn.answer_to_question.parse(call.data)
-    abiturient_question = call.message.text.split("\n")[-1]
+    callback_data: dict = group_btn.answer_to_question.parse(call.data)
+    matriculant_question: str = call.message.text.split("\n")[-1]
 
     await call.message.delete_reply_markup()
 
     async with state.proxy() as data:
         data["questioner_id"] = callback_data["user_id"]
-        data["question"] = abiturient_question
+        data["question"] = matriculant_question
 
     await call.message.answer("Напишите ответ")
     await Questions.get_answer.set()
@@ -150,7 +155,4 @@ async def help_button(message: Message):
         target_user=message.chat.type
     ).first().content
 
-    await message.answer(
-        content or "В этот раздел ещё не добавили информацию"
-    )
-# -------------------------------------
+    await message.answer(content or "В этот раздел ещё не добавили информацию")
