+++ handlers/users/__init__.py
@@ -1,3 +1,5 @@
+from .questions import dp
+from .inner_menu import dp
 from .main_menu import dp
 from .btn_back_handler import dp
 from .commands import dp
