+++ utils/__init__.py
@@ -1,2 +1,1 @@
-from . import db_api
-from . import google_sheets
+from . import db_api, google_sheets, scheduler, bot_commands
